realtime_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2

# Local NLP
torch==2.4.1
transformers==4.44.2

# Caching
cachetools==5.5.0
